seprate modal from main component
